requirements.txt
tensorflow
torch
ultralytics
datasets
pillow
opencv-python
streamlit
matplotlib
numpy
tqdm
